Variants Load and Variants Transform changed to be components with JobScope.
Removed duplicated Code in VariantsLoad and VariantsTransform to generate URL

Code to generate URL refactore to a util class.
